updated heroku_msa syntax

+++ heroku_msa.sql
@@ -1,7 +1,7 @@
 CREATE TABLE msa_exp (
   geoid int,
-  lon numeric,
-  lat numeric
+  lon double precision,
+  lat double precision
   );
 
 \copy msa_exp FROM 'msa.csv' DELIMITER ',' CSV HEADER;
